requirements.txt
Django==3.0.5
psycopg2-binary==2.8.5
requests==2.23.0
toml==0.10.0
python-telegram-bot==12.6.1